--- packages/server/src/http/internals/HTTPHandlerOptions.ts
@@ -1,34 +1,0 @@
-import {
-  AnyRouter,
-  inferRouterContext,
-  inferRouterError,
-  ProcedureType,
-} from '../../router';
-import { TRPCResponse } from '../../rpc';
-import { TRPCError } from '../../TRPCError';
-import { ResponseMeta } from '../ResponseMeta';
-import { BaseHandlerOptions } from '../../internals/BaseHandlerOptions';
-
-type ResponseMetaFn<TRouter extends AnyRouter> = (opts: {
-  data: TRPCResponse<unknown, inferRouterError<TRouter>>[];
-  ctx?: inferRouterContext<TRouter>;
-  /**
-   * The different tRPC paths requested
-   **/
-  paths?: string[];
-  type: ProcedureType | 'unknown';
-  errors: TRPCError[];
-}) => ResponseMeta;
-
-/**
- * Base interface for anything using HTTP
- */
-export interface HTTPBaseHandlerOptions<TRouter extends AnyRouter, TRequest>
-  extends BaseHandlerOptions<TRouter, TRequest> {
-  /**
-   * Add handler to be called before response is sent to the user
-   * Useful for setting cache headers
-   * @link https://trpc.io/docs/caching
-   */
-  responseMeta?: ResponseMetaFn<TRouter>;
-}
