+++ packages/server/src/http/internals/resolveHttpResponse.ts
@@ -12,8 +12,12 @@ import {
 import { TRPCErrorResponse, TRPCResponse, TRPCResultResponse } from '../../rpc';
 import { TRPCError } from '../../TRPCError';
 import { getHTTPStatusCode } from './getHTTPStatusCode';
-import { HTTPBaseHandlerOptions } from './HTTPHandlerOptions';
-import { HTTPHeaders, HTTPRequest, HTTPResponse } from './types';
+import {
+  HTTPHeaders,
+  HTTPRequest,
+  HTTPResponse,
+  HTTPBaseHandlerOptions,
+} from './types';
 
 const HTTP_METHOD_PROCEDURE_TYPE_MAP: Record<
   string,
