+++ packages/server/src/http/internals/types.ts
@@ -1,5 +1,15 @@
-import { Dict } from '@trpc/server';
+import {
+  AnyRouter,
+  Dict,
+  inferRouterContext,
+  inferRouterError,
+  ProcedureType,
+  ResponseMeta,
+  TRPCError,
+} from '@trpc/server';
 import { URLSearchParams } from 'url';
+import { BaseHandlerOptions } from '../../internals/BaseHandlerOptions';
+import { TRPCResponse } from '../../rpc';
 
 export type HTTPHeaders = Dict<string | string[]>;
 
@@ -14,4 +24,28 @@ export interface HTTPRequest {
   query: URLSearchParams;
   headers: HTTPHeaders;
   body: unknown;
+}
+
+type ResponseMetaFn<TRouter extends AnyRouter> = (opts: {
+  data: TRPCResponse<unknown, inferRouterError<TRouter>>[];
+  ctx?: inferRouterContext<TRouter>;
+  /**
+   * The different tRPC paths requested
+   **/
+  paths?: string[];
+  type: ProcedureType | 'unknown';
+  errors: TRPCError[];
+}) => ResponseMeta;
+
+/**
+ * Base interface for anything using HTTP
+ */
+export interface HTTPBaseHandlerOptions<TRouter extends AnyRouter, TRequest>
+  extends BaseHandlerOptions<TRouter, TRequest> {
+  /**
+   * Add handler to be called before response is sent to the user
+   * Useful for setting cache headers
+   * @link https://trpc.io/docs/caching
+   */
+  responseMeta?: ResponseMetaFn<TRouter>;
 }
